refactor(LottoPurchaseMoney): 패키지 구조 및 클래스명 변경, 테스트 케이스 이름 수정
- Money 클래스명을 LottoPurchaseMoney로 변경하여 로또 구입 금액과 관련된 역할이 명확히 드러나도록 개선함
- 패키지 구조 및 테스트 케이스명을 좀 더 구분하기 쉽고 일관되게 변경함
