+++ src/test/java/lotto/domain/LottoPurchaseMoneyTest.java
@@ -7,54 +7,54 @@ import static camp.nextstep.edu.missionutils.test.Assertions.assertSimpleTest;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-class MoneyTest extends NsTest {
+class LottoPurchaseMoneyTest extends NsTest {
     private static final String ERROR_MESSAGE = "[ERROR]";
 
     @Test
-    void 실패_금액이_1000원_미만이면_예외가_발생한다() {
+    void 구입금액이_1000원_미만이면_예외가_발생한다() {
         assertSimpleTest(() -> {
             IllegalArgumentException exception = assertThrows(IllegalArgumentException.class, () -> {
-                new Money("500");
+                new LottoPurchaseMoney("500");
             });
             assertThat(exception.getMessage()).contains(ERROR_MESSAGE);
         });
     }
 
     @Test
-    void 실패_금액이_1000원_단위가_아니면_예외가_발생한다() {
+    void 구입금액이_1000원_단위가_아니면_예외가_발생한다() {
         assertSimpleTest(() -> {
             IllegalArgumentException exception = assertThrows(IllegalArgumentException.class, () -> {
-                new Money("1500");
+                new LottoPurchaseMoney("1500");
             });
             assertThat(exception.getMessage()).contains(ERROR_MESSAGE);
         });
     }
 
     @Test
-    void 실패_금액이_int_범위를_넘으면_예외가_발생한다() {
+    void 구입금액이_int_범위를_넘으면_예외가_발생한다() {
         assertSimpleTest(() -> {
             IllegalArgumentException exception = assertThrows(IllegalArgumentException.class, () -> {
-                new Money("2147483648"); // int 범위를 넘는 값
+                new LottoPurchaseMoney("2147483648"); // int 범위를 넘는 값
             });
             assertThat(exception.getMessage()).contains(ERROR_MESSAGE);
         });
     }
 
     @Test
-    void 실패_금액이_숫자가_아니면_예외가_발생한다() {
+    void 구입금액이_숫자가_아니면_예외가_발생한다() {
         assertSimpleTest(() -> {
             IllegalArgumentException exception = assertThrows(IllegalArgumentException.class, () -> {
-                new Money("abc");
+                new LottoPurchaseMoney("abc");
             });
             assertThat(exception.getMessage()).contains(ERROR_MESSAGE);
         });
     }
 
     @Test
-    void 실패_금액이_공백이면_예외가_발생한다() {
+    void 구입금액이_공백이면_예외가_발생한다() {
         assertSimpleTest(() -> {
             IllegalArgumentException exception = assertThrows(IllegalArgumentException.class, () -> {
-                new Money("");
+                new LottoPurchaseMoney("");
             });
             assertThat(exception.getMessage()).contains(ERROR_MESSAGE);
         });
